style: add and apply ruff.toml
This commit adds a `ruff.toml` configuration to the project as
a precursor to using a `pyproject.toml`.

The configuration is based on pyuvdata's ruff section of their
pyproject.toml with site-specific options removed.

The configuration was then applied to the source with:

```bash
ruff format demo/
ruff check --fix demo/
```
The resulting automated changes are also included in this commit.